Move methods to the commandUtils
