+++ src/java/org/apache/cassandra/tools/nodetool/CommandUtils.java
@@ -20,20 +20,31 @@ package org.apache.cassandra.tools.nodetool;
 
 import java.io.PrintStream;
 import java.lang.reflect.Field;
+import java.net.UnknownHostException;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
+import java.util.Collections;
 import java.util.Comparator;
 import java.util.List;
+import java.util.Map;
 import java.util.Set;
+import java.util.SortedMap;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
+import com.google.common.collect.Maps;
+
+import org.apache.cassandra.locator.EndpointSnitchInfoMBean;
+import org.apache.cassandra.tools.NodeProbe;
 import org.apache.cassandra.tools.nodetool.formatter.TableBuilder;
 import org.apache.cassandra.tools.nodetool.layout.CassandraUsage;
 import org.apache.cassandra.utils.Pair;
 
+import static com.google.common.collect.Iterables.toArray;
 import static java.util.stream.Stream.concat;
 import static java.util.stream.Stream.ofNullable;
+import static org.apache.commons.lang3.ArrayUtils.EMPTY_STRING_ARRAY;
 
 /**
  * Utility methods for nodetool commands.
@@ -138,5 +149,81 @@ public final class CommandUtils
             table.add(value);
 
         table.printTo(out);
+    }
+
+    public static List<String> parseOptionalKeyspace(List<String> cmdArgs, NodeProbe nodeProbe)
+    {
+        return parseOptionalKeyspace(cmdArgs, nodeProbe, KeyspaceSet.ALL);
+    }
+
+    public static List<String> parseOptionalKeyspaceNonLocal(List<String> cmdArgs, NodeProbe nodeProbe)
+    {
+        return parseOptionalKeyspace(cmdArgs, nodeProbe, KeyspaceSet.NON_LOCAL_STRATEGY);
+    }
+
+    private static List<String> parseOptionalKeyspace(List<String> cmdArgs, NodeProbe nodeProbe, KeyspaceSet defaultKeyspaceSet)
+    {
+        List<String> keyspaces = new ArrayList<>();
+
+
+        if (cmdArgs == null || cmdArgs.isEmpty())
+        {
+            if (defaultKeyspaceSet == KeyspaceSet.NON_LOCAL_STRATEGY)
+                keyspaces.addAll(keyspaces = nodeProbe.getNonLocalStrategyKeyspaces());
+            else if (defaultKeyspaceSet == KeyspaceSet.NON_SYSTEM)
+                keyspaces.addAll(keyspaces = nodeProbe.getNonSystemKeyspaces());
+            else
+                keyspaces.addAll(nodeProbe.getKeyspaces());
+        }
+        else
+        {
+            keyspaces.add(cmdArgs.get(0));
+        }
+
+        for (String keyspace : keyspaces)
+        {
+            if (!nodeProbe.getKeyspaces().contains(keyspace))
+                throw new IllegalArgumentException("Keyspace [" + keyspace + "] does not exist.");
+        }
+
+        return Collections.unmodifiableList(keyspaces);
+    }
+
+    public static String[] parseOptionalTables(List<String> cmdArgs)
+    {
+        return cmdArgs.size() <= 1 ? EMPTY_STRING_ARRAY : toArray(cmdArgs.subList(1, cmdArgs.size()), String.class);
+    }
+
+    public static String[] parsePartitionKeys(List<String> cmdArgs)
+    {
+        return cmdArgs.size() <= 2 ? EMPTY_STRING_ARRAY : toArray(cmdArgs.subList(2, cmdArgs.size()), String.class);
+    }
+
+    public static SortedMap<String, SetHostStatWithPort> getOwnershipByDcWithPort(NodeProbe probe, boolean resolveIp,
+                                                                                  Map<String, String> tokenToEndpoint,
+                                                                                  Map<String, Float> ownerships)
+    {
+        SortedMap<String, SetHostStatWithPort> ownershipByDc = Maps.newTreeMap();
+        EndpointSnitchInfoMBean epSnitchInfo = probe.getEndpointSnitchInfoProxy();
+        try
+        {
+            for (Map.Entry<String, String> tokenAndEndPoint : tokenToEndpoint.entrySet())
+            {
+                String dc = epSnitchInfo.getDatacenter(tokenAndEndPoint.getValue());
+                if (!ownershipByDc.containsKey(dc))
+                    ownershipByDc.put(dc, new SetHostStatWithPort(resolveIp));
+                ownershipByDc.get(dc).add(tokenAndEndPoint.getKey(), tokenAndEndPoint.getValue(), ownerships);
+            }
+        }
+        catch (UnknownHostException e)
+        {
+            throw new RuntimeException(e);
+        }
+        return ownershipByDc;
+    }
+
+    private enum KeyspaceSet
+    {
+        ALL, NON_SYSTEM, NON_LOCAL_STRATEGY
     }
 }
