+++ src/java/org/apache/cassandra/tools/nodetool/DescribeCluster.java
@@ -29,7 +29,6 @@ import org.apache.cassandra.locator.DynamicEndpointSnitch;
 import org.apache.cassandra.locator.InetAddressAndPort;
 import org.apache.cassandra.locator.LocationInfoMBean;
 import org.apache.cassandra.tools.NodeProbe;
-import org.apache.cassandra.tools.NodeTool;
 import picocli.CommandLine.Command;
 
 @Command(name = "describecluster", description = "Print the name, snitch, partitioner and schema version of a cluster")
@@ -113,7 +112,7 @@ public class DescribeCluster extends WithPortDisplayAbstractCommand
             System.exit(1);
         }
 
-        SortedMap<String, SetHostStatWithPort> dcs = NodeTool.getOwnershipByDcWithPort(probe, resolveIp, tokensToEndpoints, ownerships);
+        SortedMap<String, SetHostStatWithPort> dcs = CommandUtils.getOwnershipByDcWithPort(probe, resolveIp, tokensToEndpoints, ownerships);
 
         out.println("\nData Centers: ");
         for (Map.Entry<String, SetHostStatWithPort> dc : dcs.entrySet())
