+++ src/java/org/apache/cassandra/tools/nodetool/DisableAutoCompaction.java
@@ -26,8 +26,8 @@ import org.apache.cassandra.tools.nodetool.layout.CassandraUsage;
 import picocli.CommandLine.Command;
 import picocli.CommandLine.Parameters;
 
-import static org.apache.cassandra.tools.NodeTool.NodeToolCmd.parseOptionalKeyspace;
-import static org.apache.cassandra.tools.NodeTool.NodeToolCmd.parseOptionalTables;
+import static org.apache.cassandra.tools.nodetool.CommandUtils.parseOptionalKeyspace;
+import static org.apache.cassandra.tools.nodetool.CommandUtils.parseOptionalTables;
 import static org.apache.cassandra.tools.nodetool.CommandUtils.concatArgs;
 
 @Command(name = "disableautocompaction", description = "Disable autocompaction for the given keyspace and table")
